Revert "Поменял везде типы продаж с int на long"
This reverts commit 882305b9b456f67791bb2acfb92c0cef3b31affd.

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        long[] sales = new long[]{1,2,300,200};
+        int[] sales = new int[]{1,2,300,200};
         SalesManager salesManager = new SalesManager(sales);
-        long max = salesManager.max();
+        int max = salesManager.max();
         System.out.println(max);
 
     }
